requirements.txt
crewai>=0.11.0
gradio>=3.50.0
python-dotenv>=1.0.0
duckduckgo-search>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=2.0.0 